Consolidate test helpers and OpenBao constants
Move wiremock and fake docker helpers into tests/support and reuse them
in integration tests. Group OpenBao paths, roles, and TTL constants into
a dedicated init submodule.

Closes #74
